Pull runChromaticBuild out to separate file
